All sokoban examples seem to be working.
It's still pretty slow. The current implementation
doesn't quickly detect obvious dead ends,
such as a 2x2 of diamonds. Also, it doesn't understand
that when there are many paths to get from A to B
(that don't involve moving any blocks), you only
need to pick one of the paths. So it works well
on tightly packed problems, but it doesn't work well
for problems with wide open spaces (ironic, I know).